Integrate checkstyle functionality
