+++ src/main/java/mambo/command/ListCommand.java
@@ -1,0 +1,30 @@
+package mambo.command;
+
+import mambo.MamboException;
+import mambo.TaskListFile;
+import mambo.Ui;
+import mambo.task.TaskList;
+
+/**
+ * Represents a single "list" command that has been passed into the chatbot.
+ *
+ * @author kentalim2
+ */
+public class ListCommand extends Command {
+
+    public ListCommand(String argument) {
+        super(argument);
+    }
+
+    /**
+     * Displays the current list of tasks to the user
+     *
+     * @param tasks Task List that is being tracked by chatbot
+     * @param file Saved local file containing tasks
+     * @throws MamboException Throws exception non-proper format is used to add task
+     */
+    @Override
+    public void execute(Ui ui, TaskList tasks, TaskListFile file) {
+        System.out.println(ui.respond("these are your tasks!\n" + tasks));
+    }
+}
