+++ text-ui-test/EXPECTED.TXT
@@ -1,6 +1,6 @@
 
 ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
-hi! I'm Mambo, your personal assistant and chatbot!
+hi! I'm mambo, your personal assistant and chatbot!
 what can I do for you today? ei, ei mun!
 ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
 List of commands:
